requirements.txt
asgiref==3.4.1
Django==3.2.6
django-environ==0.4.5
django-extensions==3.1.3
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
django-cors-headers==3.7.0
djangorestframework